Backend/requirements.txt
alembic==1.13.1
asgiref==3.7.2
bleach==6.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
filelock==3.13.3
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
Markdown==3.6
MarkupSafe==2.1.5
packaging==24.0
platformdirs==4.2.0
psycopg2==2.9.9
pyswip==0.2.10
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.4
virtualenv==20.25.1
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.0.2
WTForms==3.1.2
zipp==3.18.1
